Decompose src/agents/router.rs into modular routing components
- Split 799-line router.rs into focused modules under src/agents/routing/
- Created assignment.rs (128 lines) for branch operations and agent assignment
- Created filters.rs (159 lines) for issue filtering and routing logic
- Created decisions.rs (51 lines) for priority and routing decisions
- Created coordination.rs (221 lines) for multi-agent coordination
- Created mod.rs (8 lines) for module organization
- Reduced main router.rs from 799 to 298 lines (63% reduction)
- Preserved all existing functionality and public API
- Maintained atomic operations and error handling patterns
- All modules follow clear separation of concerns
- Tests updated to work with new modular structure

Fixes #155 - Prevents router.rs from exceeding 800-line limit
