+++ src/agents/routing/mod.rs
@@ -1,0 +1,9 @@
+pub mod assignment;
+pub mod filters;
+pub mod decisions;
+pub mod coordination;
+
+pub use assignment::*;
+pub use filters::*;
+pub use decisions::*;
+pub use coordination::*;
